feat: seperate module by function
- torch template에서 제공하는 형식에 맞추어 모듈을 나누었다.
- 고려사항: model의 sentimentclassifier를 어떻게 추상화 하는게에 대한 고려가 필요하다.

- issue 확인
 #5
